requirements.txt
defusedxml==0.6.0
dmc2gym @ git+https://github.com/c-rizz/dmc2gym
empy==3.3.4
imageio==2.33.1
lr-gym==0.0.1
nptyping==1.3.0
numpy-quaternion==2022.4.2
opencv-python==4.5.5.62
pip-chill==1.0.3
pybullet==3.2.5
pygame==2.5.2
pynvml==11.0.0
pyvirtualdisplay==2.1
quaternionic==1.0.5
rospkg==1.3.0
stable-baselines3==2.2.1
tensorboard==2.14.0
torchvision==0.15.1
vidgear==0.3.0
wandb==0.16.1
xacro==1.13.3
xmltodict==0.13.0
